update client-side file structure
