+++ client/components/auth/styles.js
@@ -1,0 +1,8 @@
+const styles = {
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+  margin: "0em auto 15em",
+};
+
+export default styles;
