+++ client/components/data/styles.js
@@ -1,0 +1,17 @@
+const styles = {
+  all: {
+    // width: "100%",
+    height: "100vh",
+    backgroundColor: "#e9e9e9",
+    flexGrow: 1,
+  },
+  container: {
+    padding: "1em 0em 1em 1em",
+    minWidth: "80%",
+  },
+  menuColumn: {
+    zIndex: 1,
+  },
+};
+
+export default styles;
