+++ client/components/games/styles.js
@@ -1,0 +1,93 @@
+const styles = {
+  all: {
+    width: "100%",
+    backgroundColor: "#e9e9e9",
+    flexGrow: 1,
+  },
+  container: {
+    padding: "2em 0em 4em",
+    minWidth: "65%",
+  },
+  leftContainer: {
+    backgroundColor: "#ffffff",
+    padding: "3em 2em",
+    marginBottom: "4em",
+    borderRadius: "10px",
+    boxShadow: "0 0 25px rgba(0,0,0,.04)",
+    textAlign: "center",
+  },
+  rightContainerTop: {
+    backgroundColor: "#ffffff",
+    maxHeight: "380px",
+    minWidth: "330px",
+    padding: "2em",
+    marginBottom: "3em",
+    borderRadius: "10px",
+    boxShadow: "0 0 25px rgba(0,0,0,.04)",
+    textAlign: "center",
+  },
+  rightContainerBot: {
+    backgroundColor: "#ffffff",
+    minWidth: "330px",
+    padding: "2em",
+    marginTop: "-2em",
+    marginBottom: "3em",
+    borderRadius: "10px",
+    boxShadow: "0 0 25px rgba(0,0,0,.04)",
+    textAlign: "center",
+  },
+  title: {
+    fontSize: '4em',
+    marginTop: '0.5em',
+    fontFamily: "Fredoka One",
+    fontWeight: 500,
+  },
+  subTitle: {
+    fontSize: '2em',
+    fontFamily: "Fredoka One",
+    fontWeight: 200,
+  },
+  gameGroup: {
+    margin: '2.5em auto 1em',
+  },
+  buttonCol: {
+    margin: '0.6em auto',
+  },
+  gameContainer: {
+    marginTop: '0.6em',
+  },
+  gameButton: {
+    height: '100%',
+    width: '90%',
+    background: '#2b282e',
+    borderRadius: '20px',
+    boxShadow: " 0px 8px 5px -5px rgba(0,0,0,0.4)",
+  },
+  gameIcon: {
+    margin: 'auto',
+    color: '#ffd600',
+  },
+  gameLabel: {
+    margin: '0.8em auto 0 auto',
+    color: '#ffd600',
+  },
+  levelContainer: {
+    marginTop: '6em',
+    marginBottom: '-5em',
+  },
+  levelIcon: {
+    fontSize: '11em',
+    display: 'inline-block',
+    position: 'relative',
+    color: '#ffd600',
+  },
+  levelNumber: {
+    fontWeight: 100,
+    fontSize: '0.6em',
+  },
+  userCol: {
+    minWidth: '8em',
+  },
+};
+
+export default styles;
