+++ client/components/games/synonym-game/styles.js
@@ -1,0 +1,44 @@
+const styles = {
+  title: {
+    paddingTop: "6rem",
+    fontSize: 20,
+  },
+  titleContainer: {
+    padding: 50,
+    display: "flex",
+    justifyContent: "space-between",
+    color: "#ffffff",
+  },
+  button: {
+    height: 100,
+    width: 100,
+    borderRadius: "50%",
+    backgroundColor: '#2b282e',
+    cursor: "pointer",
+    paddingTop: 40,
+    fontSize: 20,
+    boxShadow: "0px 0px 7px rgba(0,0,0,0.1)",
+  },
+  div: {
+    backgroundColor: "#ffd600",
+    textAlign: "center",
+    height: "100vh",
+  },
+  timer: {
+    color: "#2b282e",
+    fontSize: 30,
+  },
+  scores: {
+    width: 200,
+  },
+  h3: {
+    fontSize: 20,
+  },
+  gameOver: {
+    fontFamily: "Fredoka One",
+    fontSize: 62,
+    color: "#ffffff",
+  },
+};
+
+export default styles;
