+++ client/components/main/styles.js
@@ -1,0 +1,15 @@
+const styles = {
+  body: {
+    display: 'flex',
+    minHeight: '100vh',
+    flexDirection: 'column',
+  },
+  containerFull: {
+    marginTop: '6em',
+    flex: '1',
+    display: 'flex',
+    flexDirection: 'column',
+  },
+};
+
+export default styles;
