+++ client/components/navbar/styles.js
@@ -1,0 +1,28 @@
+const styles = {
+  navbar: {
+    height: "6rem",
+    background: '#2b282e',
+  },
+  logo: {
+    paddingTop: 15,
+    paddingLeft: 20,
+    paddingBottom: 5,
+    paddingRight: 5,
+  },
+  menuItem: {
+    color: '#ffffff',
+    fontWeight: 'bold',
+  },
+  menuMenu: {
+    fontSize: '20px',
+    color: '#ffffff',
+  },
+  rightMenu: {
+    fontSize: '20px',
+    marginRight: 30,
+    color: '#ffffff',
+    fontWeight: 'bold',
+  },
+};
+
+export default styles;
