+++ client/components/new-words/index.js
@@ -8,30 +8,9 @@ import { Header, Container, Menu, Icon } from 'semantic-ui-react';
 import history from '../../history';
 import LevelWordView from './LevelWords';
 import OhterUserWordView from './OtherUserWords';
+import { panel } from './styles';
 
-const styles = {
-  all: {
-    width: "100%",
-    backgroundColor: "#e9e9e9",
-    flexGrow: 1,
-  },
-  container: {
-    padding: "2em 0em 4em",
-    // minWidth: "60%",
-  },
-  subContainer: {
-    backgroundColor: "#ffffff",
-    borderRadius: "10px",
-    boxShadow: "0 0 25px rgba(0,0,0,.04)",
-    textAlign: "center",
-  },
-  iconHeader: {
-    marginTop: "1em",
-    fontSize: '2em',
-    fontFamily: "Fredoka One",
-    fontWeight: 500,
-  },
-};
+const styles = panel;
 
 class NewWordsPanel extends React.Component {
   constructor(props) {
