+++ client/components/record/index.js
@@ -2,8 +2,9 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { Icon, Button, Message } from 'semantic-ui-react';
-import { sendWords } from '../store/userWords';
+import { sendWords } from '../../store/userWords';
 import { TweenLite, Power1, TimelineMax } from 'gsap';
+import styles from './styles';
 
 const SpeechRecognition = SpeechRecognition || webkitSpeechRecognition;
 const SpeechGrammarList = SpeechGrammarList || webkitSpeechGrammarList;
@@ -143,92 +144,6 @@ class Record extends Component {
   }
 
   render() {
-    const styles = {
-      mic: {
-        margin: "auto",
-        position: "absolute",
-        top: 100,
-        left: 270,
-        cursor: "pointer",
-      },
-      outerDiv: {
-        paddingTop: 150,
-        paddingLeft: 50,
-        backgroundColor: "#ffd600",
-        height: "100vh",
-        justifyContent: "center",
-        alignItems: "center",
-        flexDirection: "column",
-      },
-      innerDiv: {
-        width: "45%",
-        float: "left",
-        position: "relative",
-      },
-      promptContainer: {
-        marginTop: "4em",
-        display: "flex",
-        alignItems: "center",
-        flexDirection: "column",
-      },
-      prompt: {
-        marginLeft: 50,
-      },
-      button: {
-        marginTop: 50,
-        marginLeft: 50,
-      },
-      instructions: {
-        color: "#ffffff",
-        weight: 500,
-        fontStyle: "italic",
-        fontFamily: "Roboto",
-        marginTop: 40,
-        marginLeft: 50,
-        marginRight: 50,
-      },
-      innerCircle: {
-        backgroundColor: "#ffffff",
-        width: 300,
-        height: 300,
-        borderRadius: "50%",
-        position: "absolute",
-        top: 25,
-        left: 175,
-        cursor: "pointer",
-        boxShadow: "0px 0px 7px rgba(0,0,0,0.1)",
-      },
-      outerCircle: {
-        width: 350,
-        height: 350,
-        border: "5px solid #ffffff",
-        borderRadius: "50%",
-        userSelect: "none",
-        position: "absolute",
-        top: 0,
-        left: 150,
-      },
-      card: {
-        backgroundColor: "#ffffff",
-        height: 250,
-        width: 400,
-        position: "absolute",
-        top: 250,
-        left: 125,
-        borderRadius: 10,
-        padding: 20,
-        textAlign: "center",
-        boxShadow: "0px 0px 7px rgba(0,0,0,0.1)",
-      },
-      link: {
-        textDecoration: "none",
-        color: "#2b282e",
-      },
-      h3: {
-        marginBottom: 11,
-      },
-    };
-
     const resultsCard = (
       <div style={styles.card}>
         <h2>
