+++ client/components/user-home/styles.js
@@ -1,0 +1,82 @@
+const styles = {
+  all: {
+    width: "100%",
+    backgroundColor: "#e9e9e9",
+    flexGrow: 1,
+  },
+  container: {
+    padding: "2em 0em 4em",
+    minWidth: "60%",
+  },
+  leftContainer: {
+    backgroundColor: "#ffffff",
+    padding: "3em 2em",
+    marginBottom: "4em",
+    borderRadius: "10px",
+    boxShadow: "0 0 25px rgba(0,0,0,.04)",
+    textAlign: "center",
+  },
+  rightContainerTop: {
+    backgroundColor: "#ffffff",
+    maxHeight: "380px",
+    minWidth: "300px",
+    padding: "2em",
+    marginBottom: "3em",
+    borderRadius: "10px",
+    boxShadow: "0 0 25px rgba(0,0,0,.04)",
+    textAlign: "center",
+  },
+  rightContainerBot: {
+    backgroundColor: "#ffffff",
+    minWidth: "300px",
+    padding: "2em",
+    marginTop: "-2em",
+    marginBottom: "3em",
+    borderRadius: "10px",
+    boxShadow: "0 0 25px rgba(0,0,0,.04)",
+    textAlign: "center",
+  },
+  userImage: {
+    margin: '3em auto',
+  },
+  title: {
+    fontSize: '3em',
+    marginTop: '0.5em',
+    fontFamily: "Fredoka One",
+    fontWeight: 500,
+  },
+  subTitle: {
+    fontSize: '2em',
+    fontFamily: "Fredoka One",
+    fontWeight: 200,
+  },
+  list: {
+    fontSize: '1.5em',
+    margin: '3em auto 2em',
+    width: '50%',
+  },
+  emailLink: {
+    color: '#2185d0',
+  },
+  editButton: {
+    marginBottom: '1em',
+    background: '#ffd600',
+    color: '#2b282e',
+  },
+  levelContainer: {
+    marginTop: '6em',
+    marginBottom: '-5em',
+  },
+  levelIcon: {
+    fontSize: '11em',
+    display: 'inline-block',
+    position: 'relative',
+    color: '#ffd600',
+  },
+  levelNumber: {
+    fontWeight: 100,
+    fontSize: '0.6em',
+  },
+};
+
+export default styles;
