+++ client/components/welcome/Splash.js
@@ -1,0 +1,22 @@
+// import React from 'react';
+// import PropTypes from 'prop-types';
+// import { Container } from 'semantic-ui-react';
+
+// const Splash = props => {
+//   const { children } = props;
+
+//   return (
+//     <Container style={styles.splash} fluid>
+//       {children}
+//     </Container>
+//   );
+// };
+
+// export default Splash;
+
+// /**
+//  * PROP TYPES
+//  */
+// Splash.propTypes = {
+//   children: PropTypes.object,
+// };
