+++ client/components/welcome/styles.js
@@ -1,0 +1,45 @@
+const styles = {
+  container: {
+    width: "50%",
+    display: "flex",
+    alignItems: "left",
+    flexWrap: "wrap",
+    flexDirection: "column",
+    marginLeft: "5rem",
+    marginTop: "15rem",
+  },
+  subContainer: {
+    display: "flex",
+    marginTop: "2rem",
+  },
+  h1: {
+    color: '#2b282e',
+    fontSize: 36,
+  },
+  h3: {
+    color: '#2b282e',
+    fontSize: 28,
+  },
+  h4: {
+    color: '#2b282e',
+    fontSize: 20,
+    margin: "auto 2rem",
+  },
+  button: {
+    marginTop: 25,
+  },
+  icon: {
+    color: '#2b282e',
+  },
+  splash: {
+    display: 'flex',
+    position: 'absolute',
+    top: 0,
+    bottom: 0,
+    background: `url('yellowLanding.jpg') no-repeat`,
+    backgroundSize: "cover",
+    marginTop: "6rem",
+  },
+};
+
+export default styles;
