+++ client/history.js
@@ -1,6 +1,6 @@
-import createHistory from 'history/createBrowserHistory'
-import createMemoryHistory from 'history/createMemoryHistory'
+import createHistory from 'history/createBrowserHistory';
+import createMemoryHistory from 'history/createMemoryHistory';
 
-const history = process.env.NODE_ENV === 'test' ? createMemoryHistory() : createHistory()
+const history = process.env.NODE_ENV === 'test' ? createMemoryHistory() : createHistory();
 
-export default history
+export default history;
